+++ src/michijs/hooks/index.ts
@@ -11,5 +11,3 @@ export { useStringTemplate } from "./useStringTemplate";
 export { useTitle } from "./useTitle";
 export { useWatch } from "./useWatch";
 export { useAsyncComputedObserve } from "./useAsyncComputedObserve";
-export { useObservePrimitive } from "./useObservePrimitive";
-export { useComputedObservePrimitive } from "./useComputedObservePrimitive";