+++ src/michijs/hooks/useObserve.ts
@@ -1,3 +1,4 @@
+import { PrimitiveValue } from "../classes/PrimitiveValue";
 import { ProxiedValue } from "../classes/ProxiedValue";
 import type { UseObserve, UseObserveInternal } from "../types";
 import { ObservableProxyHandler } from "./proxyHandlers/ObservableProxyHandler";
@@ -21,7 +22,9 @@ export const useObserveInternal: UseObserveInternal = (
  * @param item The value to be observed.
  * @returns A new observable
  */
-export const useObserve: UseObserve = (item) => {
+export const useObserve: UseObserve = (item, usePrimitive) => {
+  if (usePrimitive)
+    return new PrimitiveValue(item) as any
   const rootObservableCallback = () => result;
   const result = useObserveInternal(item, undefined, rootObservableCallback);
   return result;