+++ src/michijs/typeWards/isCSSVariable.ts
@@ -1,0 +1,4 @@
+import type { CSSVar } from "../types";
+
+export const isCSSVariable = (value: any): value is CSSVar<string> =>
+  Boolean(value?.[Symbol.toPrimitive]);